+++ Main/Features/TroopWeight/Hooks/IOnClanPartyItemUpdateProperties.cs
@@ -1,8 +1,0 @@
-using TaleWorlds.CampaignSystem.ViewModelCollection.ClanManagement;
-
-namespace TAOM.Features.TroopWeight.Hooks;
-
-public interface IOnClanPartyItemUpdateProperties
-{
-    void OnClanPartyItemUpdateProperties(ClanPartyItemVM partyItem);
-}