+++ Main/Features/TroopWeight/Hooks/IOnPartyBaseNumberOfAllMembers.cs
@@ -1,8 +1,0 @@
-using TaleWorlds.CampaignSystem.Party;
-
-namespace TAOM.Features.TroopWeight.Hooks;
-
-public interface IOnPartyBaseNumberOfAllMembers
-{
-    void OnPartyBaseNumberOfAllMembers(PartyBase partyBase, ref int __result);
-}