+++ Main/Features/TroopWeight/Hooks/IOnPartyBaseNumberOfRegularMembers.cs
@@ -1,8 +1,0 @@
-using TaleWorlds.CampaignSystem.Party;
-
-namespace TAOM.Features.TroopWeight.Hooks;
-
-public interface IOnPartyBaseNumberOfRegularMembers
-{
-    void OnPartyBaseNumberOfRegularMembers(PartyBase partyBase, ref int __result);
-}