+++ Main/Features/TroopWeight/Hooks/IOnRecruitmentVMRefreshPartyProperties.cs
@@ -1,8 +1,0 @@
-using TaleWorlds.CampaignSystem.ViewModelCollection.GameMenu.Recruitment;
-
-namespace TAOM.Features.TroopWeight.Hooks;
-
-public interface IOnRecruitmentVMRefreshPartyProperties
-{
-    void OnRecruitmentVMRefreshPartyProperties(RecruitmentVM recruitmentVM);
-}